Fix the tests for before test dependancies
Originally, before-test-suite had 1 fail and 1 pass, with the
fail test dependant on the success test but some skips were
expected in the test case.  The test was reworked to the pass test
depending on the fail test, which should cause that pass test to
be skipped.
Additionally, the implementation of the before check was corrected
to match that understanding of the implementation