Delay printing of test name until test is run
Previously, if a test ran it's dependancies, the progress display
of the dependancy would get combine with the dependant test.

The below program can show the change in behavior before and after.

(ql:quickload :fiveam)
(5am:in-suite* foo)
(5am:test bar (5am:is (= 1 1)))
(5am:test (baz :depends-on bar) (5am:is (= 1 1)))
(5am:run! 'foo)