Use Lispworks dspec functionality to locate tests by name.
End goal: I want to be able to build an emacs test runner for fiveam,
where I can M-. a failing test to jump directly to the failing test.

AFAICT, sbcl doesn't have a mechanism for custom location types. We
could potentially hack around it by definining an gentemp function
name and map back the test name to the intermediate function. But in
that case we'll need to make slime aware of this lookup possibility
too.