Avoid EVAL in the most common scenario
On Lispworks, EVAL uses interpreted code and not compiled. In
particular, if there's a bactrace LW goes into the interpreter and we
lose some lines in the backtrace.

In a followup PR I'll also add functionality to jump to the test from
the bactrace (will work only on LW for now), but I need this for that
to work.