Add WARNS check.
The SIGNALS check can correctly detect whether a block of code signals
a warning, but it also aborts the code block's execution. That's
appropriate for checking for an error, but not for a warning. This new
check confirms that a warning of a particular class is raised, a la
SIGNALS, but also allows the block of code in its scope to complete,
which may be more natural for some cases.